+++ src/main/java/com/example/demo/dto/OrderDetailDTO.java
@@ -17,12 +17,12 @@ public class OrderDetailDTO {
     @JsonProperty("order_id")
     @Min(value=1, message = "Order's ID must be > 0")
     @Schema(name = "order_id", example = "10")
-    private Long orderId;
+    private Integer orderId;
 
     @Min(value=1, message = "Product's ID must be > 0")
     @JsonProperty("product_id")
     @Schema(name = "product_id", example = "5")
-    private Long productId;
+    private Integer productId;
 
     @Min(value=0, message = "Product's ID must be >= 0")
     @Schema(example = "1299.99")
