+++ src/main/java/com/example/demo/repo/CategoryRepo.java
@@ -3,5 +3,5 @@ package com.example.demo.repo;
 import com.example.demo.models.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepo extends JpaRepository<Category, Long> {
+public interface CategoryRepo extends JpaRepository<Category, Integer> {
 }
