+++ src/main/java/com/example/demo/repo/OrderDetailRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderDetailRepo extends JpaRepository<OrderDetail, Long> {
-    List<OrderDetail> findByOrderId(Long orderId);
+public interface OrderDetailRepo extends JpaRepository<OrderDetail, Integer> {
+    List<OrderDetail> findByOrderId(Integer orderId);
 }
