+++ src/main/java/com/example/demo/repo/ProductImageRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductImageRepo extends JpaRepository<ProductImage, Long> {
-    List<ProductImage> findByProductId(Long productId);
+public interface ProductImageRepo extends JpaRepository<ProductImage, Integer> {
+    List<ProductImage> findByProductId(Integer productId);
 }
