+++ src/main/java/com/example/demo/repo/ProductRepo.java
@@ -3,7 +3,7 @@ package com.example.demo.repo;
 import com.example.demo.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepo extends JpaRepository<Product, Long> {
+public interface ProductRepo extends JpaRepository<Product, Integer> {
     // check xem sản phẩm đó với cái name đó có tồn tại không
     boolean existsByName(String name);
